requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Statistical analysis
statsmodels>=0.14.0
dcor>=0.6.0

# Graph visualization
networkx>=3.1
matplotlib>=3.7.0
graphviz>=0.20.0

# LLM API client
requests>=2.31.0
openai>=1.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

